pyproject.toml
[project]
name = "eds-researcher"
version = "0.1.0"
description = "Agentic AI research tool for EDS pain management treatments"
readme = {text = "", content-type = "text/markdown"}
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "biopython>=1.83",
    "praw>=7.7",
    "scholarly>=1.7",
    "chromadb>=0.5",
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
]

[project.scripts]
eds-researcher = "eds_researcher.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eds_researcher"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
